Create a general property-based constructor for TilingTasks
